// src/modules/wss/wss.enum.ts
export enum SocketEvent {
  ORDER_PAYMENT_EXPIRED = 'ORDER_PAYMENT_EXPIRED',
  USER_CONNECTED = 'USER_CONNECTED',
  USER_DISCONNECTED = 'USER_DISCONNECTED',
  ORDER_STATUS_UPDATED = 'ORDER_STATUS_UPDATED',
  PAYMENT_SUCCESSFUL = 'PAYMENT_SUCCESSFUL',

  // Chat
  JOIN_CONVERSATION = 'JOIN_CONVERSATION',
  SEND_MESSAGE = 'SEND_MESSAGE',
  NEW_MESSAGE = 'NEW_MESSAGE',
  NEW_CONVERSATION = 'NEW_CONVERSATION',
  JOIN_ADMIN = 'JOIN_ADMIN',

  //PRODUCT
  PRODUCT_PRICE_CHANGED = 'PRODUCT_PRICE_CHANGED',
}